api/requirements.txt
asgiref==3.5.0
Django==4.0.3
django-filter==21.1
djangorestframework==3.13.1
mysqlclient==2.1.0
PyJWT==2.3.0
pytz==2022.1
sqlparse==0.4.2
django-cors-headers
gunicorn